Implement the second pass with size limit
It will first try to encode just with the set --crf, and see if the
size limit won't be surpassed. If it is surpassed, it will try to calculate
the bitrate for the size limit. If it fails with the calculated bitrate,
it will be recalculated removing the percentage that was surpassed until
the right size is obtained.